OutboundMessagePool and MessageRetryService
When OutboundMessagePool is unable to send a message to a peer, it
forwards it to the MessageRetryService. The service will mark the
OutboundMessage as failed and add it to an internal BinaryHeap.
OutboundMessage implements ordering based on how far it is from being
scheduled to send. Once a message is scheduled to send, it will send the
message back to the OutboundMessagePool send connection.

Each failure to send results in an exponentially increasing backoff time
in being requeued for sending, until `max_attempts` is reached at which
point the message is discarded.

Most of this services time is spent sleeping (wake on message or every second)
to avoid high cpu usage.